+++ src/main/kotlin/com/jetbrains/snakecharm/inspections/SmkSectionUnexpectedCallableArgsInspection.kt
@@ -6,7 +6,6 @@ import com.jetbrains.python.psi.PyArgumentList
 import com.jetbrains.python.psi.PyLambdaExpression
 import com.jetbrains.python.psi.PyReferenceExpression
 import com.jetbrains.python.psi.types.PyFunctionType
-import com.jetbrains.python.psi.types.TypeEvalContext
 import com.jetbrains.snakecharm.SnakemakeBundle
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.ALLOWED_LAMBDA_OR_CALLABLE_ARGS
 import com.jetbrains.snakecharm.lang.psi.SmkArgsSection
@@ -15,10 +14,10 @@ import com.jetbrains.snakecharm.lang.psi.SmkSubworkflowArgsSection
 
 class SmkSectionUnexpectedCallableArgsInspection : SnakemakeInspection() {
     override fun buildVisitor(
-            holder: ProblemsHolder,
-            isOnTheFly: Boolean,
-            session: LocalInspectionToolSession
-    ) = object : SnakemakeInspectionVisitor(holder, session) {
+        holder: ProblemsHolder,
+        isOnTheFly: Boolean,
+        session: LocalInspectionToolSession,
+    ) = object : SnakemakeInspectionVisitor(holder, getContext(session)) {
 
         override fun visitSmkSubworkflowArgsSection(st: SmkSubworkflowArgsSection) {
             if (st.sectionKeyword !in ALLOWED_LAMBDA_OR_CALLABLE_ARGS) {
@@ -33,20 +32,21 @@ class SmkSectionUnexpectedCallableArgsInspection : SnakemakeInspection() {
         }
 
         private fun checkArgumentList(
-                argumentList: PyArgumentList?,
-                section: SmkArgsSection
+            argumentList: PyArgumentList?,
+            section: SmkArgsSection,
         ) {
             val args = argumentList?.arguments ?: emptyArray()
             args.forEach { arg ->
                 if (arg is PyReferenceExpression && arg !is PyLambdaExpression) {
-                    val childType = TypeEvalContext.codeAnalysis(section.project, section.containingFile).getType(arg)
+                    val childType = myTypeEvalContext.getType(arg)
+
                     if (childType is PyFunctionType) {
                         registerProblem(
-                                arg,
-                                SnakemakeBundle.message(
-                                        "INSP.NAME.section.unexpected.callable.args.message",
-                                        section.sectionKeyword!!
-                                )
+                            arg,
+                            SnakemakeBundle.message(
+                                "INSP.NAME.section.unexpected.callable.args.message",
+                                section.sectionKeyword!!
+                            )
                         )
                     }
                 }