+++ src/main/kotlin/com/jetbrains/snakecharm/inspections/SmkShadowSettingsInspection.kt
@@ -8,12 +8,12 @@ import com.jetbrains.snakecharm.codeInsight.completion.ShadowSectionSettingsProv
 import com.jetbrains.snakecharm.lang.SnakemakeNames
 import com.jetbrains.snakecharm.lang.psi.SmkRuleOrCheckpointArgsSection
 
-class SmkShadowSettingsInspection : SnakemakeInspection()  {
+class SmkShadowSettingsInspection : SnakemakeInspection() {
     override fun buildVisitor(
-            holder: ProblemsHolder,
-            isOnTheFly: Boolean,
-            session: LocalInspectionToolSession
-    ) = object : SnakemakeInspectionVisitor(holder, session) {
+        holder: ProblemsHolder,
+        isOnTheFly: Boolean,
+        session: LocalInspectionToolSession,
+    ) = object : SnakemakeInspectionVisitor(holder, getContext(session)) {
 
         override fun visitSmkRuleOrCheckpointArgsSection(st: SmkRuleOrCheckpointArgsSection) {
             if (st.name != SnakemakeNames.SECTION_SHADOW) {
@@ -22,9 +22,10 @@ class SmkShadowSettingsInspection : SnakemakeInspection()  {
 
             val argument = st.argumentList?.arguments?.firstOrNull()
             if (argument is PyStringLiteralExpression &&
-                    argument.stringValue !in ShadowSectionSettingsProvider.SHADOW_SETTINGS) {
+                argument.stringValue !in ShadowSectionSettingsProvider.SHADOW_SETTINGS
+            ) {
                 registerProblem(argument.originalElement,
-                        SnakemakeBundle.message("INSP.NAME.shadow.settings"))
+                    SnakemakeBundle.message("INSP.NAME.shadow.settings"))
             }
         }
     }