+++ src/main/kotlin/com/jetbrains/snakecharm/inspections/smksl/SmkSLSubscriptionIndexOutOfBoundsInspection.kt
@@ -6,7 +6,7 @@ import com.intellij.psi.PsiElement
 import com.jetbrains.snakecharm.SnakemakeBundle
 import com.jetbrains.snakecharm.inspections.SnakemakeInspection
 import com.jetbrains.snakecharm.lang.psi.types.SmkAvailableForSubscriptionType
-import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLSubscriptionIndexKeyExpressionImpl
+import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLSubscriptionIndexKeyExpression
 import com.jetbrains.snakecharm.stringLanguage.lang.psi.references.SmkSLSubscriptionKeyReference
 
 class SmkSLSubscriptionIndexOutOfBoundsInspection : SnakemakeInspection() {
@@ -14,9 +14,9 @@ class SmkSLSubscriptionIndexOutOfBoundsInspection : SnakemakeInspection() {
         holder: ProblemsHolder,
         isOnTheFly: Boolean,
         session: LocalInspectionToolSession,
-    ) = object : SmkSLInspectionVisitor(holder, session) {
+    ) = object : SmkSLInspectionVisitor(holder, getContext(session)) {
 
-        override fun visitSmkSLSubscriptionExpressionKey(expr: SmkSLSubscriptionIndexKeyExpressionImpl) {
+        override fun visitSmkSLSubscriptionExpressionKey(expr: SmkSLSubscriptionIndexKeyExpression) {
             val ref = expr.reference
             if (ref is SmkSLSubscriptionKeyReference) {
                 val refText = ref.canonicalText