+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/SmkPsiElements.kt
@@ -7,6 +7,7 @@ import com.intellij.psi.StubBasedPsiElement
 import com.jetbrains.python.PyTokenTypes
 import com.jetbrains.python.codeInsight.controlflow.ScopeOwner
 import com.jetbrains.python.psi.*
+import com.jetbrains.python.psi.impl.PyElementImpl
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.WILDCARDS_DEFINING_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.WILDCARDS_EXPANDING_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.lang.psi.stubs.*
@@ -41,21 +42,24 @@ interface SmkUse : SmkRuleOrCheckpoint, StubBasedPsiElement<SmkUseStub> {
     fun getModuleName(): PsiElement?
 
     /**
-     * Returns an array of  [SmkReferenceExpression] which refer to overridden rules or checkpoints
+     * Returns an array of  [SmkReferenceExpression] which were defined as inherited rules
      */
-    fun getImportedRuleNames(): Array<SmkReferenceExpression>?
+    fun getDefinedReferencesOfImportedRuleNames(): Array<SmkReferenceExpression>?
 
     /**
-     * Returns True if it uses '*' pattern in section, where inherited rules are defined
+     * Returns list of [SmkRuleOrCheckpoint] from defined module. Returns null if there are no module reference, module section or file
      */
-    fun hasPatternInDefinitionOfInheritedRules(): Boolean
+    fun getImportedRules(): List<SmkRuleOrCheckpoint>?
 
     /**
-     * Returns the first leaf of element it is a USE_NAME_IDENTIFIER, otherwise returns nameIdentifier.
-     * Returns null if there are no identifier
+     * Returns True if  "as" name is wildcard with '*', e.g:
+     * use rule A,B from M as new_*
      */
-    fun getIdentifierLeaf(): PsiElement?
+    fun nameIdentifierIsWildcard(): Boolean
 }
+
+class SmkUseNameIdentifier(node: ASTNode) : PyElementImpl(node)
+class SmkImportedRulesNames(node: ASTNode) : PyElementImpl(node)
 
 interface SmkRuleOrCheckpointArgsSection : SmkArgsSection, PyTypedElement { // PyNamedElementContainer
     /**