+++ src/test/kotlin/com/jetbrains/python/PythonMockSdk.kt
@@ -36,6 +36,7 @@ import java.io.File
  *
  * @author yole
  */
+//
 object PythonMockSdk {
     fun create(
         testDataRoot: String,